requirements.txt
fastapi
uvicorn
youtube-transcript-api
google-genai
python-dotenv
pydantic[email]
PyMuPDF
extract-msg
python-docx
beautifulsoup4
requests
yt-dlp
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator
psycopg2-binary
bcrypt
